+++ git_machete/generated_docs.py
@@ -244,7 +244,11 @@ long_docs: Dict[str, str] = {
 
         <b>Git config keys</b>
 
-           `machete.github.{remote,domain,organization,repository}`
+           `machete.github.{domain,remote,organization,repository}`
+             `machete.github.domain`
+                The domain of the GitHub API server, for use with GitHub Enterprise; otherwise inferred from the remote URL.
+                For example, `git config machete.github.domain git.example.org`
+
              `machete.github.remote`
                 The name of the git remote (as in `git remote`) that git-machete pushes the head branch to.
                 Unless both `machete.github.organization` and `machete.github.repository` are set, this remote's URL is also inspected
@@ -253,10 +257,6 @@ long_docs: Dict[str, str] = {
                 By default (when this key is unset), if exactly one remote's URL corresponds to GitHub, that remote is selected automatically;
                 set this key to disambiguate when more than one remote points to GitHub.
                 For example, `git config machete.github.remote origin`
-
-             `machete.github.domain`
-                The domain of the GitHub API server, for use with GitHub Enterprise; otherwise inferred from the remote URL.
-                For example, `git config machete.github.domain git.example.org`
 
              `machete.github.organization`
                 The GitHub organization (the part before `/` in `organization/repository`); otherwise inferred from the remote URL.
@@ -289,7 +289,11 @@ long_docs: Dict[str, str] = {
               * `up-only-no-branches` — same as `up-only`, but no branch names are included (only PR numbers & titles)
               * `none`                — prepend no intro to the PR description at all
 
-           `machete.gitlab.{remote,domain,namespace,project}`
+           `machete.gitlab.{domain,remote,namespace,project}`
+             `machete.gitlab.domain`
+                The domain of the GitLab API server, for use with a GitLab self-managed instance; otherwise inferred from the remote URL.
+                For example, `git config machete.gitlab.domain git.example.org`
+
              `machete.gitlab.remote`
                 The name of the git remote (as in `git remote`) that git-machete pushes the source branch to.
                 Unless both `machete.gitlab.namespace` and `machete.gitlab.project` are set, this remote's URL is also inspected
@@ -298,10 +302,6 @@ long_docs: Dict[str, str] = {
                 By default (when this key is unset), if exactly one remote's URL corresponds to GitLab, that remote is selected automatically;
                 set this key to disambiguate when more than one remote points to GitLab.
                 For example, `git config machete.gitlab.remote origin`
-
-             `machete.gitlab.domain`
-                The domain of the GitLab API server, for use with a GitLab self-managed instance; otherwise inferred from the remote URL.
-                For example, `git config machete.gitlab.domain git.example.org`
 
              `machete.gitlab.namespace`
                 The GitLab namespace (the part before the final `/` in `namespace/project`); otherwise inferred from the remote URL.
@@ -789,7 +789,11 @@ long_docs: Dict[str, str] = {
 
         <b>Git config keys</b>
 
-           `machete.github.{remote,domain,organization,repository}` (all subcommands)
+           `machete.github.{domain,remote,organization,repository}` (all subcommands)
+             `machete.github.domain`
+                The domain of the GitHub API server, for use with GitHub Enterprise; otherwise inferred from the remote URL.
+                For example, `git config machete.github.domain git.example.org`
+
              `machete.github.remote`
                 The name of the git remote (as in `git remote`) that git-machete pushes the head branch to.
                 Unless both `machete.github.organization` and `machete.github.repository` are set, this remote's URL is also inspected
@@ -798,10 +802,6 @@ long_docs: Dict[str, str] = {
                 By default (when this key is unset), if exactly one remote's URL corresponds to GitHub, that remote is selected automatically;
                 set this key to disambiguate when more than one remote points to GitHub.
                 For example, `git config machete.github.remote origin`
-
-             `machete.github.domain`
-                The domain of the GitHub API server, for use with GitHub Enterprise; otherwise inferred from the remote URL.
-                For example, `git config machete.github.domain git.example.org`
 
              `machete.github.organization`
                 The GitHub organization (the part before `/` in `organization/repository`); otherwise inferred from the remote URL.
@@ -988,7 +988,11 @@ long_docs: Dict[str, str] = {
 
         <b>Git config keys</b>
 
-           `machete.gitlab.{remote,domain,namespace,project}` (all subcommands)
+           `machete.gitlab.{domain,remote,namespace,project}` (all subcommands)
+             `machete.gitlab.domain`
+                The domain of the GitLab API server, for use with a GitLab self-managed instance; otherwise inferred from the remote URL.
+                For example, `git config machete.gitlab.domain git.example.org`
+
              `machete.gitlab.remote`
                 The name of the git remote (as in `git remote`) that git-machete pushes the source branch to.
                 Unless both `machete.gitlab.namespace` and `machete.gitlab.project` are set, this remote's URL is also inspected
@@ -997,10 +1001,6 @@ long_docs: Dict[str, str] = {
                 By default (when this key is unset), if exactly one remote's URL corresponds to GitLab, that remote is selected automatically;
                 set this key to disambiguate when more than one remote points to GitLab.
                 For example, `git config machete.gitlab.remote origin`
-
-             `machete.gitlab.domain`
-                The domain of the GitLab API server, for use with a GitLab self-managed instance; otherwise inferred from the remote URL.
-                For example, `git config machete.gitlab.domain git.example.org`
 
              `machete.gitlab.namespace`
                 The GitLab namespace (the part before the final `/` in `namespace/project`); otherwise inferred from the remote URL.